letsquiz_backend/requirements.txt
# Production Requirements
django-environ==0.11.2
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0